# src/public/swagger/swagger.yaml
swagger: '2.0'
info:
  description: QuantSwing — Automated Intraday Options Trading Platform, Backend API
  version: 1.0.0
  title: QuantSwing Backend
basePath: /
schemes:
  - http
  - https
...
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: 'JWT access token, sent as: `Bearer <token>`'
  Token:
    type: apiKey
    name: token
    in: header
    description: 'Raw JWT access token (hedged-style `token` header). Alternative to BearerAuth.'
paths:
...
  /instruments/sync:
    get:
      operationId: syncInstrumentMasterGet
      summary: Refresh the instrument master (browser-friendly)
      description: Same as the POST variant; provided so the sync can be triggered from a browser.
      tags:
        - instrument
      produces:
        - application/json
      responses:
        200:
          description: Sync completed
          schema:
            $ref: '#/definitions/InstrumentSyncResult'
        503:
          description: Angel One scrip master unavailable
          schema:
            $ref: '#/definitions/Error'
    post:
      operationId: syncInstrumentMaster
      summary: Refresh the instrument master
      description: >
        Fetches the Angel One scrip master dump, filters it to the platform
        universe (index options on Nifty 50, Bank Nifty, and Sensex plus their
        underlying index rows), and atomically replaces the instrument table.
      tags:
        - instrument
      produces:
        - application/json
      responses:
        200:
          description: Sync completed
          schema:
            $ref: '#/definitions/InstrumentSyncResult'
        503:
          description: Angel One scrip master unavailable
          schema:
            $ref: '#/definitions/Error'
  /instruments/search:
    get:
      operationId: searchInstruments
      summary: Search instruments
      tags:
        - instrument
      produces:
        - application/json
      parameters:
        - name: q
          in: query
          required: false
          type: string
          description: Case-insensitive substring match on the trading symbol
        - name: name
          in: query
          required: false
          type: string
          description: Underlying name (NIFTY, BANKNIFTY, SENSEX)
        - name: instrumentType
          in: query
          required: false
          type: string
          enum:
            - OPTIDX
            - AMXIDX
        - name: expiry
          in: query
          required: false
          type: string
          format: date
        - name: limit
          in: query
          required: false
          type: integer
          default: 50
          maximum: 200
      responses:
        200:
          description: Matching instruments
          schema:
            type: array
            items:
              $ref: '#/definitions/Instrument'
        401:
          description: Unauthenticated
          schema:
            $ref: '#/definitions/Error'
  /instruments/expiries:
    get:
      operationId: getInstrumentExpiries
      summary: Distinct upcoming expiries for an underlying
      tags:
        - instrument
      produces:
        - application/json
      parameters:
        - name: name
          in: query
          required: true
          type: string
          description: Underlying name (NIFTY, BANKNIFTY, SENSEX)
      responses:
        200:
          description: Upcoming expiry dates, ascending
          schema:
            type: array
            items:
              type: string
              format: date
        401:
          description: Unauthenticated
          schema:
            $ref: '#/definitions/Error'
  /instruments/subscribe:
    post:
      operationId: subscribeInstrumentLtp
      summary: Add an instrument to the live LTP feed
      description: >
        Subscribes the instrument's token on the market-data WebSocket so
        `ltp_update` socket frames start covering it (index instruments are
        subscribed at boot; traded instruments are auto-subscribed at order
        placement). Idempotent.
      tags:
        - instrument
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/SubscribeLtpRequest'
      responses:
        200:
          description: Subscription outcome
          schema:
            $ref: '#/definitions/SubscribeLtpResponse'
        404:
          description: Instrument not found
          schema:
            $ref: '#/definitions/Error'
  /orders:
    post:
      operationId: placeOrder
      summary: Place an order
      description: >
        Persists the order as PENDING and enqueues it for placement with the
        broker. The transition to OPEN/REJECTED arrives asynchronously (poll
        the order or listen for `order-update` socket events). MARKET orders
        are resolved to a LIMIT at the current price before placement.
      tags:
        - order
      consumes:
        - application/json
      produces:
        - application/json
      security:
        - BearerAuth: []
        - Token: []
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/PlaceOrderRequest'
      responses:
        201:
          description: Order created (PENDING)
          schema:
            $ref: '#/definitions/Order'
        400:
          description: Invalid payload, lot-size/freeze-qty violation, or no broker session
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Instrument not found
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthenticated
          schema:
            $ref: '#/definitions/Error'
    get:
      operationId: listOrders
      summary: List orders (newest first)
      tags:
        - order
      produces:
        - application/json
      security:
        - BearerAuth: []
        - Token: []
      parameters:
        - name: status
          in: query
          required: false
          type: string
          enum:
            - PENDING
            - OPEN
            - PARTIAL
            - COMPLETED
            - CANCELLED
            - REJECTED
        - name: limit
          in: query
          required: false
          type: integer
          default: 50
          maximum: 200
      responses:
        200:
          description: Orders with their instrument
          schema:
            type: array
            items:
              $ref: '#/definitions/Order'
        401:
          description: Unauthenticated
          schema:
            $ref: '#/definitions/Error'
  /orders/{id}:
    get:
      operationId: getOrder
      summary: Get one order with its recent broker logs
      tags:
        - order
      produces:
        - application/json
      security:
        - BearerAuth: []
        - Token: []
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        200:
          description: The order, its instrument, and up to 20 recent broker logs
          schema:
            $ref: '#/definitions/Order'
        404:
          description: Order not found
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthenticated
          schema:
            $ref: '#/definitions/Error'
    patch:
      operationId: modifyOrder
      summary: Modify an OPEN order
      description: >
        Persists the new price/quantity and enqueues a broker modification.
        The outcome (fill or no-op) arrives asynchronously.
      tags:
        - order
      consumes:
        - application/json
      produces:
        - application/json
      security:
        - BearerAuth: []
        - Token: []
      parameters:
        - name: id
          in: path
          required: true
          type: string
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/ModifyOrderRequest'
      responses:
        200:
          description: Updated order
          schema:
            $ref: '#/definitions/Order'
        409:
          description: Order is not OPEN
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Order not found
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthenticated
          schema:
            $ref: '#/definitions/Error'
  /orders/{id}/cancel:
    post:
      operationId: cancelOrder
      summary: Request cancellation of an OPEN order
      tags:
        - order
      produces:
        - application/json
      security:
        - BearerAuth: []
        - Token: []
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        202:
          description: Cancellation enqueued
          schema:
            $ref: '#/definitions/MessageResponse'
        409:
          description: Order is not OPEN
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Order not found
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthenticated
          schema:
            $ref: '#/definitions/Error'
...
  /positions/{id}/orders:
    post:
      operationId: placePositionOrders
      summary: Place a batch of diff orders against the position's legs
      description: >
        The legacy place_order flow — the client computes the required orders
        from held vs target quantities (sync) or reverse orders (exit) and
        posts them together. Orders carrying an `id` modify that OPEN order.
        Placement is risk-sorted (exits/hedges first). Market hours only.
      tags: [position]
      consumes: [application/json]
      produces: [application/json]
      security:
        - BearerAuth: []
        - Token: []
      parameters:
        - name: id
          in: path
          required: true
          type: string
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/PlacePositionOrdersRequest'
      responses:
        201:
          description: Persisted orders in placement order
          schema:
            type: object
            properties:
              orders:
                type: array
                items:
                  $ref: '#/definitions/Order'
        400:
          description: Market closed / validation failure
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Position closed or order no longer OPEN
          schema:
            $ref: '#/definitions/Error'
  /orders/{id}/recheck:
    post:
      operationId: recheckOrder
      summary: Force a broker status check for an order
      tags: [order]
      produces: [application/json]
      security:
        - BearerAuth: []
        - Token: []
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        202:
          description: Status check enqueued
          schema:
            $ref: '#/definitions/MessageResponse'
        404:
          description: Order not found
          schema:
            $ref: '#/definitions/Error'
...
definitions:
  Instrument:
    type: object
    title: Instrument
    description: One tradable/watchable contract from the filtered Angel One scrip master.
    properties:
      id:
        type: string
        description: '`${exchSeg}:${symbol}`'
        example: 'NFO:NIFTY29JAN2621000CE'
      token:
        type: string
        description: Angel One instrument token (used for the WS feed and orders)
      symbol:
        type: string
        example: NIFTY29JAN2621000CE
      name:
        type: string
        description: Underlying name
        example: NIFTY
      expiry:
        type: string
        format: date
        x-nullable: true
      strike:
        type: number
        example: 21000
      lotSize:
        type: integer
      tickSize:
        type: number
      instrumentType:
        type: string
        enum:
          - OPTIDX
          - AMXIDX
      exchSeg:
        type: string
        enum:
          - NFO
          - BFO
          - NSE
          - BSE
      freezeQty:
        type: integer
      lastPrice:
        type: number
        description: Last traded price, refreshed by the LTP stream
      volume:
        type: number
        description: Cumulative day volume, refreshed by the LTP stream
      createdAt:
        type: string
        format: date-time
      updatedAt:
        type: string
        format: date-time
  InstrumentSyncResult:
    type: object
    title: InstrumentSyncResult
    description: Outcome of one instrument master refresh.
    properties:
      fetched:
        type: integer
        description: Rows fetched from the Angel One dump before filtering
      stored:
        type: integer
        description: Rows stored after applying the universe filter
...
  MessageResponse:
    type: object
    title: MessageResponse
    properties:
      message:
        type: string
        example: Logged out
...
  PositionOrderInput:
    type: object
    title: PositionOrderInput
    required: [positionLegId, transactionType, orderType, quantity]
    properties:
      id:
        type: string
        description: Present — modify this OPEN order instead of creating one
      positionLegId:
        type: string
      transactionType:
        type: string
        enum: [BUY, SELL]
      orderType:
        type: string
        enum: [MARKET, LIMIT]
      quantity:
        type: integer
      price:
        type: number
        description: Required for LIMIT
      productType:
        type: string
        enum: [INTRADAY, NORMAL]
        default: INTRADAY
  PlacePositionOrdersRequest:
    type: object
    title: PlacePositionOrdersRequest
    required: [orders]
    properties:
      orders:
        type: array
        items:
          $ref: '#/definitions/PositionOrderInput'
  SubscribeLtpRequest:
    type: object
    title: SubscribeLtpRequest
    required:
      - instrumentId
    properties:
      instrumentId:
        type: string
        example: 'NFO:NIFTY21JUL2624100CE'
  SubscribeLtpResponse:
    type: object
    title: SubscribeLtpResponse
    properties:
      subscribed:
        type: boolean
        description: False when the LTP stream is offline (no broker credentials)
      message:
        type: string
        example: Subscribed to live LTP
  Order:
    type: object
    title: Order
    description: >
      A single order routed to a broker. Status flow: PENDING (queued) -> OPEN
      (accepted) -> COMPLETED/CANCELLED/REJECTED. `order-update` socket events
      fire on every transition.
    properties:
      id:
        type: string
        format: uuid
      userId:
        type: string
        format: uuid
      instrumentId:
        type: string
        example: 'NFO:NIFTY29JAN2621000CE'
      instrument:
        $ref: '#/definitions/Instrument'
      broker:
        type: string
        enum:
          - PAPER
      status:
        type: string
        enum:
          - PENDING
          - OPEN
          - PARTIAL
          - COMPLETED
          - CANCELLED
          - REJECTED
      transactionType:
        type: string
        enum:
          - BUY
          - SELL
      orderType:
        type: string
        description: MARKET orders are resolved to LIMIT at placement time
        enum:
          - MARKET
          - LIMIT
      productType:
        type: string
        enum:
          - INTRADAY
          - NORMAL
      quantity:
        type: integer
      price:
        type: number
        description: Limit price
      brokerOrderId:
        type: string
        x-nullable: true
        description: Broker-side order id; null until accepted
        example: PAPER-8fa85a5d-a2b2-4b1f-899e-5014436056ef
      filledQuantity:
        type: integer
      averageExecutionPrice:
        type: number
        x-nullable: true
      rejectReason:
        type: string
        x-nullable: true
      placedAt:
        type: string
        format: date-time
        x-nullable: true
      executedAt:
        type: string
        format: date-time
        x-nullable: true
      cancelledAt:
        type: string
        format: date-time
        x-nullable: true
      brokerLogs:
        type: array
        description: Present only on GET /orders/{id} (up to 20 recent)
        items:
          $ref: '#/definitions/BrokerLog'
      positionId:
        type: string
        x-nullable: true
        description: OMS linkage; null for quick-ticket orders
      positionLegId:
        type: string
        x-nullable: true
      stt:
        type: number
        x-nullable: true
      stampDuty:
        type: number
        x-nullable: true
      exchangeCharges:
        type: number
        x-nullable: true
      sebiCharges:
        type: number
        x-nullable: true
      brokerage:
        type: number
        x-nullable: true
      gst:
        type: number
        x-nullable: true
      totalCharges:
        type: number
        x-nullable: true
        description: Sum of all charges, stamped on COMPLETED fills
      chase:
        type: boolean
        description: Whether the price chase re-quotes this order while OPEN
      chaseAttempts:
        type: integer
      chaseFailures:
        type: integer
      remarks:
        type: string
        x-nullable: true
      createdAt:
        type: string
        format: date-time
      updatedAt:
        type: string
        format: date-time
  BrokerLog:
    type: object
    title: BrokerLog
    description: Audit trail of one broker API call (real or simulated).
    properties:
      id:
        type: string
        format: uuid
      broker:
        type: string
        enum:
          - PAPER
      method:
        type: string
        example: POST
      url:
        type: string
        example: https://paper.test.broker/place-order
      request:
        type: object
        x-nullable: true
      response:
        type: object
        x-nullable: true
      isError:
        type: boolean
      statusCode:
        type: integer
      duration:
        type: integer
        description: Call duration in ms
      orderId:
        type: string
        x-nullable: true
      brokerOrderId:
        type: string
        x-nullable: true
      userId:
        type: string
        x-nullable: true
      createdAt:
        type: string
        format: date-time
  PlaceOrderRequest:
    type: object
    title: PlaceOrderRequest
    required:
      - instrumentId
      - transactionType
      - orderType
      - quantity
    properties:
      instrumentId:
        type: string
        example: 'NFO:NIFTY29JAN2621000CE'
      transactionType:
        type: string
        enum:
          - BUY
          - SELL
      orderType:
        type: string
        enum:
          - MARKET
          - LIMIT
      quantity:
        type: integer
        description: Must be a multiple of the instrument lot size
      price:
        type: number
        description: Required for LIMIT orders
      productType:
        type: string
        enum:
          - INTRADAY
          - NORMAL
        default: INTRADAY
  ModifyOrderRequest:
    type: object
    title: ModifyOrderRequest
    description: At least one of price or quantity is required.
    properties:
      price:
        type: number
      quantity:
        type: integer
...
  Error:
    type: object
    title: Error
    description: Standard error envelope returned by the central error handler.
    properties:
      error:
        type: object
        properties:
          code:
            type: string
            example: UNAUTHORIZED
          message:
            type: string
            example: Invalid credentials
          requestId:
            type: string
          details:
            type: object
            x-nullable: true
          stack:
            type: string
            description: Present only outside production.